feat: pss-cln-work-contract-roots-002 - Fold content staging and materialization root helpers into private subservices
