+++ internal/ownershipinventory/work_root_contract_test.go
@@ -31,6 +31,7 @@ func TestWorkThinRootContractFiles(t *testing.T) {
 		"content_contract.go",
 		"content_materialization_public_seam_test.go",
 		"content_materialize_contract.go",
+		"content_staging_contract.go",
 		"content_staging_public_seam_test.go",
 		"contracts.go",
 		"input.go",
@@ -67,12 +68,10 @@ func TestWorkExcessRootContractFoldDestinations(t *testing.T) {
 	t.Parallel()
 
 	want := map[string]string{
-		"request_admission":            "pkg/services/work/internal",
-		"invocation_return_policy":     "pkg/services/work/internal",
-		"lineage_graph_modules":        "pkg/services/work/internal/services/state_access",
-		"state_access_query":           "pkg/services/work/internal/services/state_access",
-		"content_staging_impl":         "pkg/services/work/internal/services/content_staging",
-		"content_materialization_impl": "pkg/services/work/internal/services/content_materialization",
+		"request_admission":        "pkg/services/work/internal",
+		"invocation_return_policy": "pkg/services/work/internal",
+		"lineage_graph_modules":    "pkg/services/work/internal/services/state_access",
+		"state_access_query":       "pkg/services/work/internal/services/state_access",
 	}
 
 	for _, target := range ownershipinventory.WorkExcessRootContractFolds {
@@ -100,8 +99,6 @@ func TestWorkExcessRootContractFoldDestinations(t *testing.T) {
 	}
 	slices.Sort(gotClusters)
 	wantClusters := []string{
-		"content_materialization_impl",
-		"content_staging_impl",
 		"invocation_return_policy",
 		"lineage_graph_modules",
 		"request_admission",
