+++ pkg/services/work/internal/services/content_materialization/internal/service/materialize.go
@@ -7,7 +7,8 @@ import (
 	"strings"
 	"time"
 
-	work "github.com/portpowered/infinite-you/pkg/services/work"
+	"github.com/portpowered/infinite-you/pkg/services/work"
+	"github.com/portpowered/infinite-you/pkg/services/work/internal/contenturl"
 	contentmaterialization "github.com/portpowered/infinite-you/pkg/services/work/internal/services/content_materialization"
 )
 
@@ -133,7 +134,7 @@ func (o *Options) allowPrivateURLs() bool {
 // For http(s) and data URLs a bounded temp file is created; callers must invoke cleanup when done.
 func MaterializeContentURL(ctx context.Context, rawURL string, opts *Options) (localPath string, cleanup CleanupFunc, err error) {
 	trimmed := strings.TrimSpace(rawURL)
-	if err := work.ValidateContentURL(trimmed); err != nil {
+	if err := contenturl.Validate(trimmed); err != nil {
 		return "", noopCleanup, fmt.Errorf("scheme not supported: %s", trimmed)
 	}
